fix(decisions): ensure taxonomy terms are created when saving categories
The new updateDecisionInstance flow saves categories to config but
never created the corresponding taxonomy terms. This caused
getProcessCategories to return empty results and broke category
display in the proposals list and category duplication.

Extract ensureProposalTaxonomy to a shared utility and call it from
updateDecisionInstance when config.categories changes.
